# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReconFlow AI",
    "description": "Delivery-platform reconciliation for restaurants.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/deployment/license": {
      "post": {
        "tags": [
          "deployment"
        ],
        "summary": "Record License",
        "operationId": "record_license_deployment_license_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordLicenseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/deployment/updates": {
      "get": {
        "tags": [
          "deployment"
        ],
        "summary": "List Updates",
        "operationId": "list_updates_deployment_updates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UpdateEventResponse"
                  },
                  "type": "array",
                  "title": "Response List Updates Deployment Updates Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "deployment"
        ],
        "summary": "Record Update",
        "operationId": "record_update_deployment_updates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analyses/{analysis_id}/ai-insights": {
      "post": {
        "tags": [
          "ai-insights"
        ],
        "summary": "Generate Ai Insight",
        "operationId": "generate_ai_insight_analyses__analysis_id__ai_insights_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIInsightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses": {
      "post": {
        "tags": [
          "analysis-orchestration"
        ],
        "summary": "Create Analysis",
        "operationId": "create_analysis_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analysis-orchestration"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/versions": {
      "post": {
        "tags": [
          "analysis-orchestration"
        ],
        "summary": "Create New Version",
        "operationId": "create_new_version_analyses__analysis_id__versions_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/mark-processing": {
      "post": {
        "tags": [
          "analysis-orchestration"
        ],
        "summary": "Mark Processing",
        "operationId": "mark_processing_analyses__analysis_id__mark_processing_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/mark-completed": {
      "post": {
        "tags": [
          "analysis-orchestration"
        ],
        "summary": "Mark Completed",
        "operationId": "mark_completed_analyses__analysis_id__mark_completed_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/mark-failed": {
      "post": {
        "tags": [
          "analysis-orchestration"
        ],
        "summary": "Mark Failed",
        "operationId": "mark_failed_analyses__analysis_id__mark_failed_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkFailedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIInsightResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "executive_summary": {
            "type": "string",
            "title": "Executive Summary"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_id",
          "executive_summary",
          "provider_name",
          "model_name",
          "generated_at"
        ],
        "title": "AIInsightResponse"
      },
      "AnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "branch_id": {
            "type": "string",
            "title": "Branch Id"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "parent_analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Analysis Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "branch_id",
          "created_by",
          "parent_analysis_id",
          "version",
          "status",
          "period_start",
          "period_end",
          "failure_reason",
          "created_at",
          "updated_at"
        ],
        "title": "AnalysisResponse"
      },
      "AnalysisStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "AWAITING_FILES",
          "PROCESSING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "AnalysisStatus"
      },
...
      "CreateAnalysisRequest": {
        "properties": {
          "branch_id": {
            "type": "string",
            "title": "Branch Id"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          }
        },
        "type": "object",
        "required": [
          "branch_id",
          "period_start",
          "period_end"
        ],
        "title": "CreateAnalysisRequest"
      },
...
      "DeploymentInfoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "current_version": {
            "type": "string",
            "title": "Current Version"
          },
          "license_status": {
            "$ref": "#/components/schemas/LicenseStatus"
          },
          "license_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Expires At"
          },
          "installed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Installed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "current_version",
          "license_status",
          "license_expires_at",
          "installed_at"
        ],
        "title": "DeploymentInfoResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LicenseStatus": {
        "type": "string",
        "enum": [
          "UNLICENSED",
          "ACTIVE",
          "EXPIRED"
        ],
        "title": "LicenseStatus"
      },
...
      "MarkFailedRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "MarkFailedRequest"
      },
...
      "RecordLicenseRequest": {
        "properties": {
          "license_key": {
            "type": "string",
            "title": "License Key"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "license_key"
        ],
        "title": "RecordLicenseRequest"
      },
      "RecordUpdateRequest": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "RecordUpdateRequest"
      },
...
      "UpdateEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "applied_at": {
            "type": "string",
            "format": "date-time",
            "title": "Applied At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "applied_at",
          "notes"
        ],
        "title": "UpdateEventResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}